requirements.txt
Django==1.3.1
django-redis==1.4.5
PIL==1.1.7
django-tagging
django-mptt
